+++ gc_meox_tms/__main__.py
@@ -2,7 +2,7 @@ import argparse
 import sys
 
 from concurrent.futures import ProcessPoolExecutor
-from .gc_meox_tms import process_one_mol
+from .derivatization import process_one_mol
 from .utils import read_input_txt, write_flat, write_tab_separated
